+++ src/_internals/item/Item.tsx
@@ -89,6 +89,22 @@ export const Item = (props: ItemProps) => {
   const a11yAttrs = pickA11yProps<ItemProps>(props)
   let Tag = tag.type
   let tagProps = tag.props
+
+  // The code below make sure that if we use a HTML button, a proper type='button' or
+  // type='submit' is added. type='button' is the default.
+  // This type=button is needed to make sure non-submit buttons are not activated when
+  // submitting forms.
+  // See: https://stackoverflow.com/questions/4763638/enter-triggers-button-click
+  let tagTypeProp = {}
+  const isButtonTag = tag.type === 'button' && typeof href !== 'string'
+  if (isButtonTag) {
+    if (tag.props.type) {
+      tagTypeProp = { type: tag.props.type }
+    } else {
+      tagTypeProp = { type: 'button' }
+    }
+  }
+
   if (href) {
     if (typeof href !== 'string') {
       Tag = href.type
@@ -123,6 +139,7 @@ export const Item = (props: ItemProps) => {
   return (
     <Tag
       {...tagProps}
+      {...tagTypeProp}
       onClick={onClick}
       onFocus={onFocus}
       onBlur={onBlur}